Added in hold left, hold right and fixed error with I piece
There was an issue in ClearedLinesCheck, causing an outof bounds error if the I piece was dropped to the floor. Also added in holdleft and holdright commands, although the speed at which the piece moves will need to be adjusted. Still having issues with the game randomly stopping on start.